#147-transfer ui items into dedicated package
Intermediate commit. texts translations, refactored logo image display and adaptive text.
