+++ ArcheryTrainingTimer/app/src/main/java/com/github/schmouk/archerytrainingtimer/ui/commons/AdaptiveText.kt
@@ -1,0 +1,61 @@
+package com.github.schmouk.archerytrainingtimer.ui.commons
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.TextUnit
+import androidx.compose.ui.unit.sp
+
+
+/**
+ * AdaptiveText composable - Notice: actually unused, maybe erroneous!
+ */
+@Composable
+fun AdaptiveText(
+    text: String,
+    modifier: Modifier = Modifier,
+    color: Color = Color.Unspecified,
+    fontWeight: FontWeight? = null,
+    fontStyle: FontStyle? = null,
+    targetWidth: Dp,
+    maxLines: Int = 1,
+    initialFontSize: TextUnit = 80.sp
+) {
+    //var textSize by remember(text, targetWidth) { mutableStateOf(initialFontSize) }
+    //var readyToDraw by remember { mutableStateOf(false) }
+    var textSize = initialFontSize
+    var readyToDraw = false
+
+    Text(
+        text = text,
+        modifier = modifier.then(
+            Modifier.graphicsLayer(alpha = if (readyToDraw) 1f else 0f)
+        ),
+        color = color,
+        fontWeight = fontWeight,
+        fontStyle = fontStyle,
+        fontSize = textSize,
+        textAlign = TextAlign.Center,
+        maxLines = maxLines,
+        softWrap = false,
+        onTextLayout = { textLayoutResult ->
+            if (textLayoutResult.didOverflowWidth) {
+                if (textSize > 8.sp) { // Keep a reasonable minimum
+                    textSize *= 0.85f // Shrink a bit more aggressively if needed
+                } else {
+                    readyToDraw = true
+                }
+            } else {
+                readyToDraw = true
+            }
+        }
+    )
+}
