+++ ArcheryTrainingTimer/app/src/main/res/values-de-rDE/strings.xml
@@ -6,6 +6,7 @@
     <string name="close_app_dialog_message">Möchten Sie den Archery Training Timer wirklich beenden ?</string>
     <string name="close_app_dialog_confirm">Ausfahrt</string>
     <string name="close_app_dialog_cancel">Abbrechen</string>
+    <string name="editor_logo">Editor-Logo</string>
     <string name="empty_string"></string>
     <string name="intermediate_beeps">Zwischentöne alle 5 Sekunden</string>
     <string name="please_select">Wählen Sie Ihre Sitzungssteuerung aus :</string>
